requirements.txt
python-dotenv
langchain-pinecone
langchain-openai
langchain-core
langchain
streamlit
pandas
